+++ changelog.md
@@ -4,6 +4,7 @@ Codion Change Log
 ## 0.18.57
 ### is.codion.swing.common.ui
 - FrameBuilder.iconImage() and iconImages() added.
+- FrameBuilder.componentListener() added.
 
 ## 0.18.56
 ### is.codion.common.core